Move vendored md5 into its own private target
