+++ src/external/CMakeLists.txt
@@ -1,3 +1,4 @@
 
 add_subdirectory(catch2)
+add_subdirectory(md5)
 add_subdirectory(tomlplusplus)
