+++ src/external/md5/CMakeLists.txt
@@ -1,0 +1,6 @@
+
+add_library(VendoredMD5 STATIC md5.c)
+target_include_directories(VendoredMD5 PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>)
+
+
+
